refactor: fold testing/ into tests/
Shared mocks move to tests/mocks/ (mockstesting/ intact); the four
unit tests that lived there move to tests/unit/. Coverage include
updated; lib/db/ added so the relocated DB core stays covered.
