+++ frontend/tests/mocks/auth-mocks.ts
