+++ frontend/tests/unit/stemguid-processor-edge-cases.test.ts
